# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Backcast EVS",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/schedule-baselines": {
      "get": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Read Schedule Baselines",
        "description": "Retrieve schedule baselines with pagination.",
        "operationId": "get_schedule_baselines",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Page number (1-indexed)",
              "default": 1,
              "title": "Page"
            },
            "description": "Page number (1-indexed)"
          },
          {
            "name": "per_page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "description": "Items per page",
              "default": 20,
              "title": "Per Page"
            },
            "description": "Items per page"
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch to query",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch to query"
          },
          {
            "name": "cost_element_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter by Cost Element ID",
              "title": "Cost Element Id"
            },
            "description": "Filter by Cost Element ID"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Create Schedule Baseline",
        "description": "Create a new schedule baseline in specified branch.",
        "operationId": "create_schedule_baseline",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBaselineCreate"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleBaselineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule-baselines/{schedule_baseline_id}": {
      "get": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Read Schedule Baseline",
        "description": "Get a specific schedule baseline by id and branch.\n\nSupports time-travel queries via the as_of parameter to view\nthe baseline's state at any historical point in time.",
        "operationId": "get_schedule_baseline",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_baseline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Baseline Id"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch to query",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch to query"
          },
          {
            "name": "as_of",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Time travel: get baseline state as of this timestamp (ISO 8601)",
              "title": "As Of"
            },
            "description": "Time travel: get baseline state as of this timestamp (ISO 8601)"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleBaselineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Update Schedule Baseline",
        "description": "Update a schedule baseline. Creates new version or forks.",
        "operationId": "update_schedule_baseline",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_baseline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Baseline Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ScheduleBaselineUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ScheduleBaselineRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Delete Schedule Baseline",
        "description": "Soft delete a schedule baseline in a branch.",
        "operationId": "delete_schedule_baseline",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_baseline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Baseline Id"
            }
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch to delete from",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch to delete from"
          },
          {
            "name": "control_date",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "date-time"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Optional control date for deletion",
              "title": "Control Date"
            },
            "description": "Optional control date for deletion"
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule-baselines/{schedule_baseline_id}/history": {
      "get": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Get Schedule Baseline History",
        "description": "Get full version history for a schedule baseline across all branches.",
        "operationId": "get_schedule_baseline_history",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_baseline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Baseline Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ScheduleBaselineRead"
                  },
                  "title": "Response Get Schedule Baseline History"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/schedule-baselines/{schedule_baseline_id}/pv": {
      "get": {
        "tags": [
          "Schedule Baselines"
        ],
        "summary": "Calculate Planned Value",
        "description": "Calculate Planned Value (PV) for a schedule baseline.\n\nPV = BAC \u00d7 Progress\n\nWhere Progress is calculated based on the baseline's progression type\n(LINEAR, GAUSSIAN, or LOGARITHMIC) and the current date relative to\nthe baseline's start and end dates.\n\nArgs:\n    schedule_baseline_id: The baseline to calculate PV for\n    current_date: The date to calculate progress for\n    bac: Budget at Completion (total planned budget)\n    branch: Branch to query (default: \"main\")\n\nReturns:\n    Dictionary with:\n    - schedule_baseline_id: The baseline ID\n    - current_date: The date used for calculation\n    - bac: Budget at Completion\n    - progress: Progress value (0.0 to 1.0)\n    - pv: Planned Value (BAC \u00d7 Progress)\n    - progression_type: Type of progression used",
        "operationId": "calculate_planned_value",
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "parameters": [
          {
            "name": "schedule_baseline_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Schedule Baseline Id"
            }
          },
          {
            "name": "current_date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "description": "Date to calculate PV for (ISO 8601)",
              "title": "Current Date"
            },
            "description": "Date to calculate PV for (ISO 8601)"
          },
          {
            "name": "bac",
            "in": "query",
            "required": true,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "string",
                  "pattern": "^(?!^[-+.]*$)[+-]?0*\\d*\\.?\\d*$"
                }
              ],
              "description": "Budget at Completion (BAC) amount",
              "title": "Bac"
            },
            "description": "Budget at Completion (BAC) amount"
          },
          {
            "name": "branch",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "description": "Branch to query",
              "default": "main",
              "title": "Branch"
            },
            "description": "Branch to query"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true,
                  "title": "Response Calculate Planned Value"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "ScheduleBaselineCreate": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Baseline name"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date",
            "description": "Schedule start date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date",
            "description": "Schedule end date"
          },
          "progression_type": {
            "type": "string",
            "title": "Progression Type",
            "description": "Type of progression curve (LINEAR, GAUSSIAN, LOGARITHMIC)",
            "default": "LINEAR"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description of the baseline"
          },
          "schedule_baseline_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schedule Baseline Id",
            "description": "Root Schedule Baseline ID (internal use only for seeding)"
          },
          "cost_element_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cost Element Id",
            "description": "ID of the cost element"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "end_date",
          "cost_element_id"
        ],
        "title": "ScheduleBaselineCreate",
        "description": "Properties required for creating a Schedule Baseline."
      },
      "ScheduleBaselineRead": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 255,
            "minLength": 1,
            "title": "Name",
            "description": "Baseline name"
          },
          "start_date": {
            "type": "string",
            "format": "date-time",
            "title": "Start Date",
            "description": "Schedule start date"
          },
          "end_date": {
            "type": "string",
            "format": "date-time",
            "title": "End Date",
            "description": "Schedule end date"
          },
          "progression_type": {
            "type": "string",
            "title": "Progression Type",
            "description": "Type of progression curve (LINEAR, GAUSSIAN, LOGARITHMIC)",
            "default": "LINEAR"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description",
            "description": "Optional description of the baseline"
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "schedule_baseline_id": {
            "type": "string",
            "format": "uuid",
            "title": "Schedule Baseline Id"
          },
          "cost_element_id": {
            "type": "string",
            "format": "uuid",
            "title": "Cost Element Id"
          },
          "created_by": {
            "type": "string",
            "format": "uuid",
            "title": "Created By"
          },
          "branch": {
            "type": "string",
            "title": "Branch"
          }
        },
        "type": "object",
        "required": [
          "name",
          "start_date",
          "end_date",
          "id",
          "schedule_baseline_id",
          "cost_element_id",
          "created_by",
          "branch"
        ],
        "title": "ScheduleBaselineRead",
        "description": "Properties returned to client."
      },
      "ScheduleBaselineUpdate": {
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string",
                "maxLength": 255,
                "minLength": 1
              },
              {
                "type": "null"
              }
            ],
            "title": "Name"
          },
          "start_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Start Date"
          },
          "end_date": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "End Date"
          },
          "progression_type": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Progression Type"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          }
        },
        "type": "object",
        "title": "ScheduleBaselineUpdate",
        "description": "Properties that can be updated on a Schedule Baseline."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "type": "oauth2",
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/auth/login"
          }
        }
      }
    }
  }
}